requirements.txt
fastapi
langchain
langchain-core
langchain-google-genai
google-cloud-aiplatform
openai 
langchain-community
langgraph 
langsmith
uvicorn 
python-dotenv
